Added Policy for getUsers and moved some APIs.
