+++ API/Controllers/UsersController.cs
@@ -24,12 +24,6 @@ namespace API.Controllers
             _userRepository = userRepository;
             _libraryRepository = libraryRepository;
         }
-
-        [HttpGet]
-        public async Task<ActionResult<IEnumerable<MemberDto>>> GetUsers()
-        {
-            return Ok(await _userRepository.GetMembersAsync());
-        }
         
         [HttpPost("add-library")]
         public async Task<ActionResult> AddLibrary(CreateLibraryDto createLibraryDto)
@@ -72,7 +66,27 @@ namespace API.Controllers
             
             return BadRequest("Not implemented");
         }
-
         
+        [Authorize(Policy = "RequireAdminRole")]
+        [HttpDelete("delete-user")]
+        public async Task<ActionResult> DeleteUser(string username)
+        {
+            var user = await _userRepository.GetUserByUsernameAsync(username);
+            _userRepository.Delete(user);
+
+            if (await _userRepository.SaveAllAsync())
+            {
+                return Ok();
+            }
+            
+            return BadRequest("Could not delete the user.");
+        }
+        
+        [Authorize(Policy = "RequireAdminRole")]
+        [HttpGet]
+        public async Task<ActionResult<IEnumerable<MemberDto>>> GetUsers()
+        {
+            return Ok(await _userRepository.GetMembersAsync());
+        }
     }
 }
